feat(runtime): coordinate configuration observation and planning

+++ IranDirect.Core/Runtime/ObservedDirectGateway.cs
@@ -8,5 +8,5 @@ public sealed record ObservedDirectGateway
 
     public required string InterfaceName { get; init; }
 
-    public int InterfaceMetric { get; init; }
+    public long InterfaceMetric { get; init; }
 }
